Establish basic brutalist CSS for index and red sauce pages
Experimented with various CSS styles & attributes to create a minimal, early web-inspired design for the homepage and the red sauce recipe page. Almost all of the CSS is in the linked style.css page, but I used a little inline styling for specific elements where it was helpful. Focused on quickly generating ideas, getting them in the code, and immediately solving problems as they came up.